+++ app/issues/[id]/EditIssueButton.tsx
@@ -1,0 +1,14 @@
+import { Pencil2Icon } from "@radix-ui/react-icons";
+import { Button } from "@radix-ui/themes";
+import Link from "next/link";
+
+const EditIssueButton = ({ issueId }: { issueId: number }) => {
+  return (
+    <Button>
+      <Pencil2Icon />
+      <Link href={`/issues/${issueId}/edit`}>Edit Issue</Link>
+    </Button>
+  );
+};
+
+export default EditIssueButton;
