+++ app/issues/[id]/IssueDetails.tsx
@@ -1,0 +1,21 @@
+import IssueStatusBadge from "@/app/components/IssueStatusBadge";
+import { Issue } from "@prisma/client";
+import { Card, Flex, Heading, Text } from "@radix-ui/themes";
+import ReactMarkdown from "react-markdown";
+
+const IssueDetails = ({ issue }: { issue: Issue }) => {
+  return (
+    <>
+      <Heading>{issue.title}</Heading>
+      <Flex gap='3' my='2'>
+        <IssueStatusBadge status={issue.status} />
+        <Text>{issue.createdAt.toDateString()}</Text>
+      </Flex>
+      <Card className='prose' mt='4'>
+        <ReactMarkdown>{issue.description}</ReactMarkdown>
+      </Card>
+    </>
+  );
+};
+
+export default IssueDetails;
