+++ app/issues/[id]/edit/page.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import IssueForm from "../../_components/IssueForm";
+import prisma from "@/prisma/client";
+import { notFound } from "next/navigation";
+
+interface Props {
+  params: { id: string };
+}
+
+const EditIssuePage = async ({ params }: Props) => {
+  const issue = await prisma.issue.findUnique({
+    where: {
+      id: parseInt(params.id)
+    }
+  });
+
+  if (!issue) {
+    notFound();
+  }
+
+  return (
+    <div>
+      <IssueForm issue={issue} />
+    </div>
+  );
+};
+
+export default EditIssuePage;
